requirements.txt
fastapi==0.88.0
sqlalchemy==1.4.45
uvicorn==0.20.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
httpx==0.23.1
pytest==7.2.0
async-sqlalchemy
greenlet